rename renderHeader to headerComponent
