+++ packages/react-admin-mui/src/MasterLayout.sc.tsx
