+++ packages/react-admin-mui/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./breadcrumbs";
 export * from "./menu";
 export * from "./ThemeProvider";
-export * from "./Master";
+export * from "./MasterLayout";
 export * from "./styled-components";
